+++ projects/04/src/util/KData.java
@@ -3,6 +3,6 @@ package util;
 import java.io.Serializable;
 
 public interface KData extends Serializable{
-	public int distanceTo(KData d);
+	public double distanceTo(KData d);
 	public String toString();
 }
